requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
jsonfield==3.1.0
psycopg2==2.9.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
